Made prediction util function. Generalised for GPU/CPU
