+++ re_train.py
@@ -12,6 +12,8 @@ import torch.nn.functional as F
 from sklearn.metrics import roc_curve
 from sklearn import metrics
 from scipy.stats import pearsonr
+from utils import predictions
+
 
 os.environ['CUDA_LAUNCH_BLOCKING'] = '1' 
 
@@ -102,19 +104,6 @@ def test(model, device, loader, criterion):
             loss += criterion(output, data.y).item()
     return loss/len(loader.dataset)
 
-def predictions(model, device, loader):
-    model.eval()
-    y_hat = torch.tensor([]).cuda()
-    y_true = torch.tensor([]).cuda()
-    with torch.no_grad():
-        for data in loader:
-            data = data.to(device)
-            output = model(data)
-            if output.dim() == 0:
-                output = output.unsqueeze(0)
-            y_hat = torch.cat((y_hat, output),0) 
-            y_true = torch.cat((y_true, data.y),0)
-    return y_hat, y_true
 
 
 # 设置训练参数
